+++ src/main/java/Shifter.java
@@ -1,6 +1,6 @@
 public class Shifter {
-    public static short[] leftShift(short[] source, int shift) {
-        short[] result = new short[source.length];
+    public static int[] leftShift(int[] source, int shift) {
+        int[] result = new int[source.length];
 
         if (source.length - 1 >= 0)
             System.arraycopy(source, 1, result, 0, source.length - 1);
